requirements.txt
asgiref==3.7.2
boto3==1.28.44
botocore==1.31.44
Django==4.2.5
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
packaging==23.1
Pillow==10.0.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
